refactor; experiment with resolution
