+++ app/feeds/show/template.hbs
@@ -1,87 +1,94 @@
-  <div class="row">
-    <div class="col-md-6">
-      <h2>{{link-to 'Feeds' 'feeds'}} / {{model.feed.onestop_id}}</h2>
-    </div>
-    <div class="col-md-6 text-right">
-      {{feed-fetcher-controls feed=model.feed}}
-    </div>
+{{#if showChangeset}}
+  {{#modal-dialog title='Changeset' ok='saveChangeset' close='hideChangeset' okText='Apply Changeset'}}
+    {{changeset-form changeset=model.changeset users=model.users }}
+  {{/modal-dialog}}
+{{/if}}
+
+<div class="row">
+  <div class="col-md-6">
+    <h2>{{link-to 'Feeds' 'feeds'}} / {{model.feed.onestop_id}}</h2>
   </div>
+  <div class="col-md-6 text-right">
+    {{feed-fetcher-controls feed=model.feed}}
+  </div>
+</div>
 
-  <dl class="dl-horizontal">
-    <dt>URL</dt>
-    <dd>{{model.feed.url}}</dd>
-    <dt>Feed format</dt>
-    <dd>{{model.feed.feed_format}}</dd>
-    <dt>Tags</dt>
-    <dd>
-      {{#each-in model.feed.tags as |key value|}}
-        {{key}}: {{value}} <br />
-      {{/each-in}}
-    </dd>
-    <dt>License name and/or URL</dt>
-    <dd>{{model.feed.license_name}} {{model.feed.license_url}}</dd>
-    <dt>License use without attribution</dt>
-    <dd>{{model.feed.license_use_without_attribution}}</dd>
-    <dt>License created derived product</dt>
-    <dd>{{model.feed.license_create_derived_product}}</dd>
-    <dt>License redistribute</dt>
-    <dd>{{model.feed.license_redistribute}}</dd>
-    {{#if model.feed.license_attribution_text}}
-      <dt>License attribution text</dt>
-      <dd>{{model.feed.license_attribution_text}}</dd>
-    {{/if}}
-    <dt># of feed versions</dt>
-    <dd>{{model.feed.feed_versions_count}}</dd>
-    <dt>Active feed version</dt>
-    <dd>
-      {{#if model.feed.active_feed_version}}
-        {{model.feed.active_feed_version.sha1}}
-      {{else}}
-        <em>none</em>
-      {{/if}}
-    </dd>
-    <dt>Import level of active feed version</dt>
-    <dd>{{number-as-pips model.feed.import_level_of_active_feed_version 4}}</dd>
-    <dt>Created</dt>
-    <dd>{{time-since-with-mouseover model.feed.created_at}}</dd>
-    <dt>Updated</dt>
-    <dd>{{time-since-with-mouseover model.feed.updated_at}}</dd>
-    {{#if model.feed.created_or_updated_in_changeset}}
-      <dt>Created or updated in</dt>
-      <dd>
-        {{#link-to 'changesets.show' model.feed.created_or_updated_in_changeset}}
-          changeset #{{model.feed.created_or_updated_in_changeset.id}}
-        {{/link-to}}
-      </dd>
-    {{/if}}
-    <dt>Last fetched at</dt>
-    <dd>{{time-since-with-mouseover model.feed.last_fetched_at}}</dd>
-    <dt>Last imported at</dt>
-    <dd>{{time-since-with-mouseover model.feed.last_imported_at}}</dd>
-  </dl>
-  {{#if model.issues}}
-    {{#each model.issues as |issue|}}
-      {{#collapsable-panel title="Error fetching feed..." panelClass="panel-danger"}}
-        <strong>Issue created by feed fetch {{moment-from-now model.feed.last_fetched_at}} ago:</strong>
-        <pre>{{#link-to 'issues.feed-fetch'}}
-          Feed Fetch Issue: #{{issue.id}}
-        {{/link-to}}</pre>
-      {{/collapsable-panel}}
-    {{/each}}
+<dl class="dl-horizontal">
+  <dt>URL</dt>
+  <dd>{{model.feed.url}}</dd>
+  <dt>Feed format</dt>
+  <dd>{{model.feed.feed_format}}</dd>
+  <dt>Tags</dt>
+  <dd>
+    {{#each-in model.feed.tags as |key value|}}
+      {{key}}: {{value}} <br />
+    {{/each-in}}
+  </dd>
+  <dt>License name and/or URL</dt>
+  <dd>{{model.feed.license_name}} {{model.feed.license_url}}</dd>
+  <dt>License use without attribution</dt>
+  <dd>{{model.feed.license_use_without_attribution}}</dd>
+  <dt>License created derived product</dt>
+  <dd>{{model.feed.license_create_derived_product}}</dd>
+  <dt>License redistribute</dt>
+  <dd>{{model.feed.license_redistribute}}</dd>
+  {{#if model.feed.license_attribution_text}}
+    <dt>License attribution text</dt>
+    <dd>{{model.feed.license_attribution_text}}</dd>
   {{/if}}
-
-  <h2>Changesets imported from this feed</h2>
-  <ul>
-    {{#each model.feed.changesets_imported_from_this_feed as |changeset|}}
-      <li>
-        {{#link-to 'changesets.show' changeset}}
-          changeset #{{changeset.id}}
-        {{/link-to}}
-      </li>
+  <dt># of feed versions</dt>
+  <dd>{{model.feed.feed_versions_count}}</dd>
+  <dt>Active feed version</dt>
+  <dd>
+    {{#if model.feed.active_feed_version}}
+      {{model.feed.active_feed_version.sha1}}
     {{else}}
-      <p>No changesets have been imported from this feed.</p>
-    {{/each}}
-  </ul>
+      <em>none</em>
+    {{/if}}
+  </dd>
+  <dt>Import level of active feed version</dt>
+  <dd>{{number-as-pips model.feed.import_level_of_active_feed_version 4}}</dd>
+  <dt>Created</dt>
+  <dd>{{time-since-with-mouseover model.feed.created_at}}</dd>
+  <dt>Updated</dt>
+  <dd>{{time-since-with-mouseover model.feed.updated_at}}</dd>
+  {{#if model.feed.created_or_updated_in_changeset}}
+    <dt>Created or updated in</dt>
+    <dd>
+      {{#link-to 'changesets.show' model.feed.created_or_updated_in_changeset}}
+        changeset #{{model.feed.created_or_updated_in_changeset.id}}
+      {{/link-to}}
+    </dd>
+  {{/if}}
+  <dt>Last fetched at</dt>
+  <dd>{{time-since-with-mouseover model.feed.last_fetched_at}}</dd>
+  <dt>Last imported at</dt>
+  <dd>{{time-since-with-mouseover model.feed.last_imported_at}}</dd>
+</dl>
+{{#if model.issues}}
+  {{#each model.issues as |issue|}}
+    {{#collapsable-panel title="Error fetching feed..." panelClass="panel-danger"}}
+      <strong>Issue created by feed fetch {{moment-from-now model.feed.last_fetched_at}} ago:</strong>
+      <pre>{{#link-to 'issues.feed-fetch.show' issue.id}}
+        Feed Fetch Issue: #{{issue.id}}
+      {{/link-to}}</pre>
+      {{#edit-buttons showChangeset="showChangeset" closeDialog="closeDialog"}}{{/edit-buttons}}
+    {{/collapsable-panel}}
+  {{/each}}
+{{/if}}
 
-  <h2>Feed versions</h2>
-  {{feed-versions-list feed=model.feed}}
+<h2>Changesets imported from this feed</h2>
+<ul>
+  {{#each model.feed.changesets_imported_from_this_feed as |changeset|}}
+    <li>
+      {{#link-to 'changesets.show' changeset}}
+        changeset #{{changeset.id}}
+      {{/link-to}}
+    </li>
+  {{else}}
+    <p>No changesets have been imported from this feed.</p>
+  {{/each}}
+</ul>
+
+<h2>Feed versions</h2>
+{{feed-versions-list feed=model.feed}}
