+++ app/issues/feed-fetch/show/controller.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Controller.extend({
+  getChanges: function() {
+    //return
+  },
+});
