+++ app/issues/feed-fetch/show/route.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model: function(params) {
+    return this.store.find('issue', params['issue_id']);
+  },
+});
