+++ app/issues/feed-fetch/show/template.hbs
@@ -1,0 +1,6 @@
+<div class="row">
+  <div class="col-md-3">
+    <p>{{model.id}}</p>
+  </div>
+  {{#edit-buttons showChangeset="showChangeset" closeDialog="closeDialog"}}{{/edit-buttons}}
+</div>
