+++ app/router.js
@@ -31,7 +31,9 @@ Router.map(function() {
     });
     this.route('new', { path: 'new/:feed_id' });
     this.route('show', { path: '/:issue_id' });
-    this.route('feed-fetch', function() {});
+    this.route('feed-fetch', function() {
+      this.route('show', { path: '/:issue_id'});
+    });
   });
   this.route('stations');
   this.route('routingcheck');
